fix: jsx column is 1 less
