+++ src/core/constants.ts
@@ -4,12 +4,7 @@ export const NodeTypes = <const>{
   ELEMENT: 1,
 };
 
-export const ElementTypes = <const>{
-  ELEMENT: 0,
-  COMPONENT: 1,
-};
-
-export const TagTypes = <number[]>[
-  ElementTypes.ELEMENT,
-  ElementTypes.COMPONENT,
+export const TagTypes = [
+  0, // ELEMENT
+  1, // COMPONENT
 ];
